updated mode operator.

+++ 0_Web programming challenges/01 Codewards/c/sample-tests/main.c
@@ -1,5 +1,5 @@
 #include <stdbool.h>
 
 bool is_even(double n) {
-    return (n == (int)n) && ((int)n % 2 == 0);
+    return (n == (long int)n) && ((long int)n % 2 == 0);
 }
